+++ crates/teloxide-core/CHANGELOG.md
@@ -12,12 +12,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New id types ([#1153](pr1153))
   - Add `PollId` struct
   - Add `CallbackQueryId` struct
+  - Add `FileId` and `FileUniqueId` structs
 
 ### Changed
 
 - New id types ([#1153](pr1153))
   - Changed `id` field to `PollId` in `PollAnswer` and in `Poll`
   - Changed `id` field to `CallbackQueryId` in `CallbackQuery` and in `answer_callback_query` method
+  - Changed `id` field to `FileId` in `File` and in `get_file` method
+  - Changed `unique_id` field to `FileUniqueId` in `File`
 
 ## 0.11.2 - 2025-04-04